+++ src/librbd/internal.h
@@ -168,7 +168,7 @@ namespace librbd {
 		       int (*cb)(uint64_t, size_t, const char *, void *),
 		       void *arg);
   int64_t diff_iterate(ImageCtx *ictx, const char *fromsnapname,
-		       uint64_t off, size_t len,
+		       uint64_t off, uint64_t len,
 		       int (*cb)(uint64_t, size_t, bool, void *),
 		       void *arg);
   ssize_t read(ImageCtx *ictx, uint64_t off, size_t len, char *buf);
